♻️ refactor: 에러 폴백 ui 수정 (Jira Ticket : CCC-48)
